requirements.txt
Flask==3.1.0
joblib==1.4.2
azure-ai-ml==1.26.4
azure-identity==1.7.0
azureml-sdk==1.45.0
pandas==2.2.3
numpy==1.23.5
requests==2.32.3
